Hibajavítások, login logika módosítás
Alapvető hibák javítva. Bent felejtett scriptek törölve. Megfelelő CSS-ek hozzáadva. Login rendszer felhasználó névre való átalakítása.

+++ src/main/java/com/example/java_gyak_beadando/Lotto_Query/LottoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LottoRepository extends JpaRepository<LottoEntity, Long> {
+public interface LottoRepository extends JpaRepository<LottoEntity, Integer> {
 
     @Query("SELECT DISTINCT h.ev FROM LottoEntity h")
     List<Integer> findDistinctYears();
